Actually fix the onSaveInstanceState crash
The problem was not my misunderstanding of the Activity lifecycle.
It was simply that if the network call has not returned, then the
Issue object would be null. Putting a guard there is the correct
fix, and so I did that.

Also, since I'm caching and re-fetching the thing anyway, and
storing just the permalink, makes more sense to just keep a
reference to the permalink, instead of the full object